move block users to account settings
